Refactor frontend silence alert use case
